[MOD] scraps

+++ scrapping/src/spider/utils.py
@@ -11,6 +11,10 @@ def folders_management(source: str):
 
 
 def download_image(url, save_path):
+    if os.path.exists(save_path):
+        print(f"L'image existe déjà : {save_path}")
+        return
+
     response = requests.get(url)
     if response.status_code == 200:
         with open(save_path, 'wb') as file:
